### 2023-08-29 #5788 @editorialbot
[PRE REVIEW]: JetNet: A Python package for accessing open datasets and benchmarking machine learning methods in high energy physics
**Submitting author:** @rkansal47 (<a href="http://orcid.org/0000-0003-2445-1060">Raghav Kansal</a>)
**Repository:** https://github.com/jet-net/JetNet
**Branch with paper.md** (empty if default branch): JOSS-2023
**Version:** v0.2.3
