### 2023-09-08 #5789 @jpata
@matthewfeickert I declare a conflict of interest, as I have joint papers in the last four years with one of the coauthors (Javier Duarte).

- https://doi.org/10.1140/epjc/s10052-021-09158-w
- http://dx.doi.org/10.1088/1742-6596/2438/1/012100
- http://cds.cern.ch/record/2842375
- and another upcoming paper related to the ones I listed above.

From my point of view, the topic in the paper under review here is sufficiently different that I would be able to offer an impartial review, but let me know what is the editorial board's view on this.
