### 2023-09-08 #5789 @matthewfeickert
> I declare a conflict of interest, as I have joint papers in the last four years with one of the coauthors (Javier Duarte).

Thanks @jpata for disclosing this. Indeed, your nice work with Javier does put you in conflict given

> As a reviewer (or editor), COIs are your present or previous association with any authors of a submission: recent (past four years) collaborators in funded research or work that is published

We appreciate you briniging this up early!
