+++ .pre-commit-config.yaml
@@ -27,7 +27,7 @@ repos:
     args: ["--profile", "black"]
 
 - repo: https://github.com/psf/black-pre-commit-mirror
-  rev: 23.7.0
+  rev: 23.9.1
   hooks:
   - id: black-jupyter
     language_version: python3
