### 2023-09-26 #5789 @matthewfeickert
> the reminder to me is probably a bug 🙂

Indeed, it is. I don't think that there's a way to remove a reminder at the moment given the [list of editorialbot commands](https://joss.readthedocs.io/en/latest/editorial_bot.html#the-complete-list-of-available-commands).
