### 2023-10-02 #5789 @smsharma
## Overall

JetNet provides an intuitive interface for accessing jet datasets and evaluating / computing losses when using them for machine learning applications in collider physics. The package has already impacted the field, having been used in several publications, and is starting to become a standard for benchmarking machine learning algorithms on a range of tasks involving structured particle physics datasets. It also lowers the barrier to entry for machine learning practitioners from adjacent fields looking to work with these kinds of datasets.

Domain-specific packages like this systematize the development and comparison of machine learning algorithms, which is super valuable, and also set an example for other domains where machine learning is starting to be widely applied. Congrats to the authors for a nice contribution, which I can easily recommend for publication in JOSS, considering the minor comments below.

## Installation and testing

I was able to install the package without problems using the instructions in the documentation. I was also able to run the unit tests using  `pytest`.

A couple suggestions in particular re: the unit test pipeline:
- Adding user-facing instructions to optionally run the tests locally (potentially for a small subset of data, and gracefully accounting for optional dependencies like `energyflow`) with `pytest`, as part of the installation instructions.
- Adding e.g. checksum-based tests to verify the integrity of downloaded datasets and re-download if they don't match. A user-side failure mode may be to interrupt the tests partway, and have subsequent tests fail due to a dataset integrity issue. This could also appear during routine usage.

## Usage

Usage instructions are provided through a combination of a Quickstart guide, API documentation, and a tutorial notebook. I was able to test out basic data loading, evals, and loss function computation using these.

A more centralized exposition of various capabilities through e.g. tutorial notebooks covering different usage aspects, cross-published in the docs, could be helpful for users looking to either use JetNet for a particular thing or to gain general familiarity, although the current documentation is certainly adequate.

## Contributions

Given the scope of the package, it is well-suited for external contributions and development milestones. More specific wording on contributions and expectations for contributing would be welcome. This could be as simple as an invitation to open an issue on GitHub for bugs or functionality requests, or a more comprehensive `CONTRIBUTING.md` file.
