### 2023-10-02 jet-net/JetNet#45 @rkansal47
Instructions to run tests locally
Suggestion from @smsharma's JOSS review: https://github.com/openjournals/joss-reviews/issues/5789#issuecomment-1743285479
