### 2023-10-02 jet-net/JetNet#46 @rkansal47
Add contribution instructions
Suggestion from @smsharma's JOSS review: https://github.com/openjournals/joss-reviews/issues/5789#issuecomment-1743285479.
