### 2023-10-02 jet-net/JetNet#47 @rkansal47
Verify integrity of existing downloaded dataset
Suggestion from @smsharma's JOSS review: https://github.com/openjournals/joss-reviews/issues/5789#issuecomment-1743285479
