### 2023-10-03 4653b01
remove walrus operator for python 3.7
