+++ .github/workflows/ci.yml
@@ -33,7 +33,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ["3.7", "3.10"]
+        python-version: ["3.7", "3.12"]
         pytest-file:
           [
             "tests/datasets/test_jetnet.py",
