### 2023-10-03 2353448
change os.system commands, print md5 statements
