### 2023-10-03 cec7c03
add download option + check integrity before loading
