+++ paper.bib
@@ -79,7 +79,10 @@
     archivePrefix = "arXiv",
     primaryClass = "hep-ph",
     reportNumber = "MIT-CTP 5519",
-    month = "1",
+    doi = "10.21468/SciPostPhys.15.4.130",
+    journal = "SciPost Phys.",
+    volume = "15",
+    pages = "130",
     year = "2023"
 }
 
@@ -115,11 +118,15 @@
 
 @article{Mikuni:2023dvk,
     author = "Mikuni, Vinicius and Nachman, Benjamin and Pettee, Mariel",
-    title = "{Fast Point Cloud Generation with Diffusion Models in High Energy Physics}",
+    title = "{Fast point cloud generation with diffusion models in high energy physics}",
     eprint = "2304.01266",
     archivePrefix = "arXiv",
     primaryClass = "hep-ph",
-    month = "4",
+    doi = "10.1103/PhysRevD.108.036025",
+    journal = "Phys. Rev. D",
+    volume = "108",
+    number = "3",
+    pages = "036025",
     year = "2023"
 }
 
