### 2023-10-09 #5789 @rkansal47
@matthewfeickert I have verified the DOIs and updated the references.
