### 2023-10-13 jet-net/JetNet#53 @matthewfeickert
[JOSS Review]: Suggested Paper Revisions
These are in reference to the [JOSS review for `JetNet`](https://github.com/openjournals/joss-reviews/issues/5789).

This is a good paper that is of appropriate length and detail, though there are some minor issues that should be addressed before publication.

While @jpata had to recuse themselves due to a conflict of interest, the [good points](https://github.com/jet-net/JetNet/issues/43) that they raised RE: describing the [state of the field](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) are still relevant. You have a good Statement of need section, but you should add a sentence about JetNet in the context of other tools in this area of HEP. While the EnergyFlow datasets are referenced, there is no discussion on the relation or contrast of JetNet's library to e.g. the [EnergyFlow library's dataset access API and functionality](https://energyflow.network/docs/datasets/).

https://github.com/jet-net/JetNet/blob/eaa365ebd23b1e609a2b2591f7185a051efe70cb/paper.md?plain=1#L34-L36

I think it is also worth noting that the datasets are hosted through Zenodo and that Zenodo should be cited.

I would additionally caution you from making a causal statement in your impact section (especially over an ill defined timeline of "months"). The relevant, and substantial, impact of `JetNet` is made clear by the extent of its use across multiple research areas.

https://github.com/jet-net/JetNet/blob/eaa365ebd23b1e609a2b2591f7185a051efe70cb/paper.md?plain=1#L44-L46
