### 2023-10-13 jet-net/JetNet#54 @matthewfeickert
[JOSS Review]: Documentation Corrections
These are in reference to the [JOSS review for `JetNet`](https://github.com/openjournals/joss-reviews/issues/5789).

There are some broken areas of the documentation that should be addressed before publication.

* There is a [broken link in the docs](https://jetnet.readthedocs.io/en/latest/#documentation) from the tutorial folder not being linked correctly

https://github.com/jet-net/JetNet/blob/86eada4f53e45a551e5372016b6b12e096eeab81/README.md?plain=1#L104-L106

* Additionally,

https://github.com/jet-net/JetNet/blob/86eada4f53e45a551e5372016b6b12e096eeab81/README.md?plain=1#L108

should probably be removed until this resource ready to be used. Only the JetNet dataset is described and the rest of the page is unfinished. The documentation presented for publication should not be in a state of unfinished development.

* While not required, as you're already using ReadTheDocs, I would also strongly suggest [enabling versioning of the documentation](https://docs.readthedocs.io/en/stable/versions.html) beyond `latest`.
