### 2023-10-13 #5789 @matthewfeickert
@rkansal47 While going through the editorial review there were some minor things that I raised as https://github.com/jet-net/JetNet/issues/53 and https://github.com/jet-net/JetNet/issues/54. Otherwise things are looking very good.

Once those have been addressed, all that's left to do now before publication is to ensure that there is a long term public archive of the code that was reviewed. In this case the code was [`JetNet` `v0.2.3`](https://github.com/jet-net/JetNet/releases/tag/v0.2.3) though there has been [some development between then and the current `HEAD`](https://github.com/jet-net/JetNet/compare/v0.2.3...86eada4f53e45a551e5372016b6b12e096eeab81), so we'd suggest making a new release (as you see fit with respect to SemVer).

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [an (optional) GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

Howerver, I think you already know this as you have a Zenodo DOI in your README (Zendo is down for upgrades as I write this, otherwise I would actually check it)!

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5598104.svg)](https://doi.org/10.5281/zenodo.5598104)

Once you make a new release of `JetNet` and have a DOI for the archive please just report them here and I'll have @editorialbot update the version listed in the review and add the archive.

Let me know if you have any questions. :+1:
