### 2023-10-16 2f0f1c6
fix readthedocs python version config
