+++ .readthedocs.yaml
@@ -19,6 +19,5 @@ sphinx:
 formats: all
 
 python:
-  version: 3.8
   install:
-    - requirements: docs/requirements.txt
+  - requirements: docs/requirements.txt
