+++ docs/requirements.txt
@@ -2,12 +2,12 @@
 
 autodocsumm
 ipykernel
-m2r2
 nbsphinx
 numpy
 readthedocs-sphinx-search
 scipy
 sphinx<7
-sphinx_rtd_theme
+sphinx_rtd_theme>=0.5.2
+m2r2
 torch
 tqdm
