+++ docs/requirements.txt
@@ -7,7 +7,7 @@ numpy
 readthedocs-sphinx-search
 scipy
 sphinx<7
-sphinx_rtd_theme>=0.5.2
+sphinx_rtd_theme==0.5.2
 m2r2
 torch
 tqdm
