+++ paper.bib
@@ -199,6 +199,21 @@
     year = "2019"
 }
 
+@article{Komiske:2019jim,
+    author        = "Komiske, Patrick T. and Mastandrea, Radha and Metodiev, Eric M. and Naik, Preksha and Thaler, Jesse",
+    title         = "{Exploring the Space of Jets with CMS Open Data}",
+    eprint        = "1908.08542",
+    archivePrefix = "arXiv",
+    primaryClass  = "hep-ph",
+    reportNumber  = "MIT-CTP 5129",
+    doi           = "10.1103/PhysRevD.101.034009",
+    journal       = "Phys. Rev. D",
+    volume        = "101",
+    number        = "3",
+    pages         = "034009",
+    year          = "2020"
+}
+
 @article{Kasieczka:2019dbj,
     author = "Butter, Anja and others",
     editor = "Kasieczka, Gregor and Plehn, Tilman",
@@ -214,15 +229,15 @@
 }
 
 @article{Kasieczka_2021,
-doi = {10.1088/1361-6633/ac36b9},
-url = {https://dx.doi.org/10.1088/1361-6633/ac36b9},
-year = {2021},
-month = {dec},
-publisher = {IOP Publishing},
-volume = {84},
-number = {12},
-pages = {124201},
-author = {Gregor Kasieczka and Benjamin Nachman and David Shih and Oz Amram and Anders Andreassen and Kees Benkendorfer and Blaz Bortolato and Gustaaf Brooijmans and Florencia Canelli and Jack H Collins and Biwei Dai and Felipe F De Freitas and Barry M Dillon and Ioan-Mihail Dinu and Zhongtian Dong and Julien Donini and Javier Duarte and D A Faroughy and Julia Gonski and Philip Harris and Alan Kahn and Jernej F Kamenik and Charanjit K Khosa and Patrick Komiske and Luc Le Pottier and Pablo Martín-Ramiro and Andrej Matevc and Eric Metodiev and Vinicius Mikuni and Christopher W Murphy and Inês Ochoa and Sang Eon Park and Maurizio Pierini and Dylan Rankin and Veronica Sanz and Nilai Sarda and Urŏ Seljak and Aleks Smolkovic and George Stein and Cristina Mantilla Suarez and Manuel Szewc and Jesse Thaler and Steven Tsan and Silviu-Marian Udrescu and Louis Vaslin and Jean-Roch Vlimant and Daniel Williams and Mikaeel Yunus},
-title = {The LHC Olympics 2020 a community challenge for anomaly detection in high energy physics},
-journal = {Reports on Progress in Physics},
+    doi = {10.1088/1361-6633/ac36b9},
+    url = {https://dx.doi.org/10.1088/1361-6633/ac36b9},
+    year = {2021},
+    month = {dec},
+    publisher = {IOP Publishing},
+    volume = {84},
+    number = {12},
+    pages = {124201},
[…]
