### 2023-10-23 jet-net/JetNet#53 @rkansal47
Thank you for the suggestions @matthewfeickert. We have incorporated most of them in #55.

Unfortunately I was not able to find a DOI or recommendation for citation for Zenodo https://github.com/zenodo/zenodo, do you have any suggestions for that?
