+++ paper.bib
@@ -241,3 +241,14 @@
     title = {The LHC Olympics 2020 a community challenge for anomaly detection in high energy physics},
     journal = {Reports on Progress in Physics},
 }
+
+@misc{Zenodo,
+  doi = {10.25495/7GXK-RD71},
+  url = {https://www.zenodo.org/},
+  author = {{European Organization For Nuclear Research} and {OpenAIRE}},
+  keywords = {FOS: Physical sciences, Publication, Dataset},
+  language = {en},
+  title = {Zenodo},
+  publisher = {CERN},
+  year = {2013}
+}
