+++ paper.bib
@@ -1,254 +1,246 @@
-@article{Chen:2021euv,
-    author = "Chen, Yifan and others",
-    title = "{A FAIR and AI-ready Higgs boson decay dataset}",
-    eprint = "2108.02214",
-    archivePrefix = "arXiv",
-    primaryClass = "hep-ex",
-    doi = "10.1038/s41597-021-01109-0",
-    month = "8",
-    year = "2022",
-    journal = "Sci. Data",
-    pages = "31",
-    volume = "9",
+@article{Buhmann:2023pmh,
+  archiveprefix = {arXiv},
+  author        = {Buhmann, Erik and Kasieczka, Gregor and Thaler, Jesse},
+  doi           = {10.21468/SciPostPhys.15.4.130},
+  eprint        = {2301.08128},
+  journal       = {SciPost Phys.},
+  pages         = {130},
+  primaryclass  = {hep-ph},
+  reportnumber  = {MIT-CTP 5519},
+  title         = {{EPiC-GAN: Equivariant Point Cloud Generation for Particle Jets}},
+  volume        = {15},
+  year          = {2023}
 }
 
 
-@incollection{NEURIPS2019_9015,
-    title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
-    author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
-    booktitle = {Advances in Neural Information Processing Systems},
-    volume = {32},
-    pages = {8024},
-    year = {2019},
-    eprint = {1912.01703},
-    archivePrefix = {arXiv},
-    publisher = {Curran Associates, Inc.},
-    url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf},
-    editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett}
+@article{Chen:2021euv,
+  archiveprefix = {arXiv},
+  author        = {Chen, Yifan and others},
+  doi           = {10.1038/s41597-021-01109-0},
+  eprint        = {2108.02214},
+  journal       = {Sci. Data},
+  pages         = {31},
+  primaryclass  = {hep-ex},
+  title         = {{A FAIR and AI-ready Higgs boson decay dataset}},
+  volume        = {9},
+  year          = {2022}
 }
 
 
 @article{Hao:2022zns,
-    author = "Hao, Zichun and Kansal, Raghav and Duarte, Javier and Chernyavskaya, Nadezda",
-    title = "{Lorentz group equivariant autoencoders}",
-    eprint = "2212.07347",
-    archivePrefix = "arXiv",
-    primaryClass = "hep-ex",
-    reportNumber = "FERMILAB-PUB-22-963-V",
-    doi = "10.1140/epjc/s10052-023-11633-5",
-    journal = "Eur. Phys. J. C",
-    volume = "83",
-    number = "6",
-    pages = "485",
-    year = "2023"
-}
-
-@inproceedings{Kansal:2021cqp,
-  archiveprefix         = {arXiv},
-  author                = {Kansal, Raghav and Duarte, Javier and Su, Hao and Orzari, Breno and Tomei, Thiago and Pierini, Maurizio and Touranakou, Mary and Vlimant, Jean-Roch and Gunopulos, Dimitrios},
[…]
