### 2023-10-26 #5789 @matthewfeickert
@rkansal47 Things look good and we can move forward with the last few steps! Can you please now make a new release of JetNet and then also deposit it in an archive with a DOI, as described in https://github.com/openjournals/joss-reviews/issues/5789#issuecomment-1761074857. If you have any questions on this step please let me know.

When you get the Zenodo archive for this release please also ensure that:
* the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
* the license listed for the archive is the same as the software license.
