### 2023-10-26 3f09286
try catch for zenodo api and move tqdm import
