### 2023-10-26 17af5c1
update examples + CI to py311 + v0.2.4
