+++ jetnet/__init__.py
@@ -7,4 +7,4 @@ import jetnet.evaluation
 import jetnet.losses
 import jetnet.utils  # noqa: F401
 
-__version__ = "0.2.3.post3"
+__version__ = "0.2.4"
