### 2023-10-26 #5789 @rkansal47
Hi @matthewfeickert, I have made a new release https://github.com/jet-net/JetNet/releases/tag/v0.2.4 and this is its DOI: https://doi.org/10.5281/zenodo.10044601. We also added @zichunhao as an author to recognize his early contributions to the library.
