### 2023-10-26 #5789 @matthewfeickert
> We also added @zichunhao as an author to recognize his early contributions to the library.

@rkansal47 that's fine, but the authors should match on the paper and the release, so can you please update the paper to reflect this?
