### 2023-10-26 #5789 @rkansal47
> > We also added @zichunhao as an author to recognize his early contributions to the library.
>
> @rkansal47 that's fine, but the authors should match on the paper and the release, so can you please update the paper to reflect this?

Yup, I believe the paper and release are both updated with this.
