### 2023-10-26 #5789 @matthewfeickert
@rkansal47 please update the metadata for the [10.5281/zenodo.10044601](https://doi.org/10.5281/zenodo.10044601) archive so that the title matches the paper title of "JetNet: A Python package for accessing open datasets and benchmarking machine learning methods in high energy physics". That's the final thing on https://github.com/openjournals/joss-reviews/issues/5789#issuecomment-1752319309, so I'll go ahead and recommend acceptacne while you take care of that! :+1:
