### 2023-10-29 6fa7c8e
Minor citation edit in JOSS paper
