### 2023-10-29 #5789 @kyleniemeyer
Hi @rkansal47, I'm doing some final checks before publishing. Can you merge the PR I made with a small edit to the paper? https://github.com/jet-net/JetNet/pull/61
